move all experimental code into partial classes
