+++ src/Microsoft.IdentityModel.Tokens/Exceptions/SecurityTokenException.cs
@@ -2,19 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Diagnostics;
 using System.Runtime.Serialization;
 
 #if NET472 || NETSTANDARD2_0 || NET6_0_OR_GREATER
 using Microsoft.IdentityModel.Logging;
 #endif
-
-using Microsoft.IdentityModel.Tokens.Experimental;
-
-#if !NET8_0_OR_GREATER
-using System.Text;
-#endif
-
 
 namespace Microsoft.IdentityModel.Tokens
 {
@@ -22,13 +14,8 @@ namespace Microsoft.IdentityModel.Tokens
     /// Represents a security token exception.
     /// </summary>
     [Serializable]
-    public class SecurityTokenException : Exception
+    public partial class SecurityTokenException : Exception
     {
-        [NonSerialized]
-        private string _stackTrace;
-
-        private ValidationError _validationError;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="SecurityTokenException"/> class.
         /// </summary>
@@ -68,44 +55,6 @@ namespace Microsoft.IdentityModel.Tokens
         protected SecurityTokenException(SerializationInfo info, StreamingContext context)
             : base(info, context)
         {
-        }
-
-        /// <summary>
-        /// Sets the <see cref="ValidationError"/> that caused the exception.
-        /// </summary>
-        /// <param name="validationError"></param>
-        internal void SetValidationError(ValidationError validationError)
-        {
-            _validationError = validationError;
-        }
-
-        /// <summary>
-        /// Gets the stack trace that is captured when the exception is created.
-        /// </summary>
-        public override string StackTrace
-        {
-            get
-            {
-                if (_stackTrace == null)
-                {
-                    if (_validationError == null)
-                        return base.StackTrace;
-#if NET8_0_OR_GREATER
-                    _stackTrace = new StackTrace(_validationError.StackFrames).ToString();
-#else
-                    StringBuilder sb = new();
-                    foreach (StackFrame frame in _validationError.StackFrames)
-                    {
-                        sb.Append(frame.ToString());
-                        sb.Append(Environment.NewLine);
-                    }
-
-                    _stackTrace = sb.ToString();
-#endif
-                }
-
-                return _stackTrace;
-            }
         }
 
         /// <summary>
