+++ src/Microsoft.IdentityModel.Xml/SignedInfo.cs
@@ -7,7 +7,7 @@ using System.IO;
 using System.Xml;
 using Microsoft.IdentityModel.Logging;
 using Microsoft.IdentityModel.Tokens;
-using Microsoft.IdentityModel.Tokens.Experimental;
+//using Microsoft.IdentityModel.Tokens.Experimental;
 using static Microsoft.IdentityModel.Logging.LogHelper;
 
 namespace Microsoft.IdentityModel.Xml
@@ -15,7 +15,7 @@ namespace Microsoft.IdentityModel.Xml
     /// <summary>
     /// Represents a XmlDsig SignedInfo element as per: https://www.w3.org/TR/2001/PR-xmldsig-core-20010820/#sec-SignedInfo
     /// </summary>
-    public class SignedInfo : DSigElement
+    public partial class SignedInfo : DSigElement
     {
         private DSigSerializer _dsigSerializer = DSigSerializer.Default;
         private string _canonicalizationMethod = SecurityAlgorithms.ExclusiveC14n;
@@ -112,41 +112,6 @@ namespace Microsoft.IdentityModel.Xml
             foreach (var reference in References)
                 reference.Verify(cryptoProviderFactory);
         }
-
-#nullable enable
-        /// <summary>
-        /// Verifies the digest of all <see cref="References"/>
-        /// </summary>
-        /// <param name="cryptoProviderFactory">supplies any required cryptographic operators.</param>
-        /// <param name="callContext"> contextual information for diagnostics.</param>
-        internal SignatureValidationError? Verify(
-            CryptoProviderFactory cryptoProviderFactory,
-#pragma warning disable CA1801
-            CallContext callContext)
-#pragma warning restore CA1801
-        {
-            if (cryptoProviderFactory == null)
-                return SignatureValidationError.NullParameter(
-                    nameof(cryptoProviderFactory),
-                    ValidationError.GetCurrentStackFrame());
-
-            SignatureValidationError? validationError = null;
-
-            for (int i = 0; i < References.Count; i++)
-            {
-                var reference = References[i];
-                validationError = reference.Verify(cryptoProviderFactory, callContext);
-
-                if (validationError is not null)
-                {
-                    validationError.AddCurrentStackFrame();
-                    break;
-                }
-            }
-
-            return validationError;
-        }
-#nullable restore
 
         /// <summary>
         /// Writes the Canonicalized bytes into a stream.
